refactor #31: 회원가입 비지니스 로직 디렉토리 수정
