--- src/mutations/useSignUpMutaion/types.ts
@@ -1,10 +1,0 @@
-export interface SignUpFormInput {
-  username: string;
-  password: string;
-  nickname: string;
-}
-
-export interface SignUp {
-  socialType: string;
-  username: string;
-}
